% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{build_ride_urls}
\alias{build_ride_urls}
\title{Build URLs to try for RIDE enrollment downloads}
\usage{
build_ride_urls(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Character vector of URLs to try
}
\description{
Build URLs to try for RIDE enrollment downloads
}
\keyword{internal}
