% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundled_data.R
\name{bundled_data_available}
\alias{bundled_data_available}
\title{Check if bundled data is available for a year}
\usage{
bundled_data_available(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
TRUE if bundled data exists for the year
}
\description{
Check if bundled data is available for a year
}
\keyword{internal}
