% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_infoworks_enrollment}
\alias{download_infoworks_enrollment}
\title{Download enrollment from InfoWorks (archived system)}
\usage{
download_infoworks_enrollment(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Data frame or NULL if download fails
}
\description{
For historical data that may only be available in the archived
InfoWorks system.
}
\keyword{internal}
