% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_ride_enrollment}
\alias{download_ride_enrollment}
\title{Download enrollment data from RIDE Data Center}
\usage{
download_ride_enrollment(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Data frame with enrollment data
}
\description{
Downloads the October 1st Public School Student Headcounts file from RIDE.
The Data Center provides Excel files with enrollment data.
}
\keyword{internal}
