% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_ride_file}
\alias{download_ride_file}
\title{Download a file from RIDE Data Center}
\usage{
download_ride_file(url, end_year)
}
\arguments{
\item{url}{URL to download from}

\item{end_year}{School year end (for temp file naming)}
}
\value{
Data frame or NULL if download fails
}
\description{
Download a file from RIDE Data Center
}
\keyword{internal}
