% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_available_years}
\alias{get_available_years}
\title{Get available years for Rhode Island enrollment data}
\usage{
get_available_years()
}
\value{
Integer vector of available end years
}
\description{
Returns the range of years available from RIDE. The RIDE Data Center
provides October 1st enrollment headcounts with data going back to 2010-11.
}
\details{
Data Eras:
\itemize{
\item Era 1 (2011-2014): Historical data with potentially different format
\item Era 2 (2015-present): Current RIDE Data Center format
}
}
\examples{
get_available_years()
}
