% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{import_local_enrollment}
\alias{import_local_enrollment}
\title{Import local enrollment file}
\usage{
import_local_enrollment(file_path, end_year)
}
\arguments{
\item{file_path}{Path to local Excel or CSV file}

\item{end_year}{School year end (for adding year column)}
}
\value{
Data frame with enrollment data
}
\description{
For cases where automated download is not possible, this function
allows importing a locally downloaded file.
}
\keyword{internal}
