% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundled_data.R
\name{load_bundled_enr}
\alias{load_bundled_enr}
\title{Load bundled enrollment data}
\usage{
load_bundled_enr(end_year)
}
\arguments{
\item{end_year}{School year end (2011-2026)}
}
\value{
Data frame with enrollment data in wide format
}
\description{
Loads the bundled October 1st enrollment data from the package.
This is the primary data source since the RIDE Data Center
requires browser-based downloads.
}
\keyword{internal}
