% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rischooldata-package.R
\docType{package}
\name{rischooldata-package}
\alias{rischooldata}
\alias{rischooldata-package}
\title{rischooldata: Rhode Island School Data}
\description{
A simple, consistent interface for accessing Rhode Island school data in Python and R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://almartin82.github.io/rischooldata}
  \item \url{https://github.com/almartin82/rischooldata}
  \item Report bugs at \url{https://github.com/almartin82/rischooldata/issues}
}

}
\author{
\strong{Maintainer}: Al Martin \email{almartin@example.com}

}
\keyword{internal}
