% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{get_raw_enr}
\alias{get_raw_enr}
\title{Download raw enrollment data from RIDE}
\usage{
get_raw_enr(end_year)
}
\arguments{
\item{end_year}{School year end (2023-24 = 2024)}
}
\value{
Data frame with raw enrollment data
}
\description{
Downloads enrollment data from RIDE's Data Center.
Uses the October 1st Public School Student Headcounts reports.
}
\keyword{internal}
