% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rischooldata-package.R
\docType{package}
\name{rischooldata-package}
\alias{rischooldata}
\alias{rischooldata-package}
\title{rischooldata: Fetch and Process Rhode Island School Data}
\description{
Downloads and processes school data from the Rhode Island Department of Education (RIDE). Provides functions for fetching enrollment data from RIDE's Data Center and transforming it into tidy format for analysis. Supports October 1st Public School Student Headcounts from 2010-2011 to present.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/almartin82/rischooldata}
  \item Report bugs at \url{https://github.com/almartin82/rischooldata/issues}
}

}
\author{
\strong{Maintainer}: Al Martin \email{almartin@example.com}

}
\keyword{internal}
